Make break_ties_core func.
Next: Rewrite break_ties_core and logic of find_global_alignment.
Here's the deal: Only once we have filled out the entire best_paths_mat,
do we have enough information to deal with ties.  For example, as we
backtrack and come to a tie, we know what it should have been.
Thus, I think we can break ties during the backtracking step.
However, this doesn't explain the current problem of the cost
not making sense for the optimal path actually chosen.