Improve variable nameing in get_next_best_costs func.
Next: get_next_best_costs is working. However, it returns a
tuple so we will have to adjust the rest of the codebase
to work with tuples later.

dp_array_forward is not working.  I don't think the list
comprehension accounts for the fact that the dp_array
arg in get_next_best_costs needs to be different each
time.